change location for reactions ans postMethod
